Added iCal proxy using existing iCal post processor.
ical.json is a new proxy configuration which allows fetching remote iCalendar
feeds. It uses the existing
org.sakaiproject.nakamura.proxy.ICalProxyPostProcessor which converts the
iCalendar feed into JSON.

This proxy needs to allow users to specify the remote host in the URL, which
requires that the post processor used is listed as safe (in the same way
that the rss post processor is). As such, the "iCal" post processor has been
added to the safe list in
org.sakaiproject.nakamura.proxy.ProxyClientServiceImpl.