Fixed proxy dropping user supplied filenames & content types for file POSTs.
When a file was included as part of a multipart form in a POST request to the
proxy, the filename= param of the Content-Disposition part header would be set
to the name of the part instead of the name of the file. e.g:

Content-Disposition: form-data; name="my_file_part"; filename="my_file_part"

This commit makes the filename= value come from the filename= value the client
sends in their POST. Additionally, the Content-Type part header uses the
mime type the client sends, rather than the default (application/octet-stream).