KERN-2680 Sorting A-Z in the Search Page All Types/Content not working
Problem was that general_sort is not set for content sakai/widget-data, resulting in their parent content being displayed out of order. Also content could appear more than once, for each widget that had content that matched the search criteria. Solution is to apply general_sort to sakai/widget-data and to group by the returnpath of the document when searching.