KERN-2723 Improve configurability of sparse database drivers.
The changes in this ticket are dependent on the changes in SPARSE-176. In that ticket, all JDBC sparse drivers are grouped into one generic JDBC driver, and they require the associated JDBC drivers to be deployed to the OSGi container as bundles.

This change provides postgresql as a deployable JDBC bundle, and updates the app list.xml to default to the generic JDBC sparse bundle, and provide commented versions of the other supported drivers commented out.