KERN-2777 Provide the ability to granularly "@Ignore"
This change allows UI devs to specify @Ignore as a key suffix in a JSON import that instructs the server to completely ignore that tree of content. No content below that node (including the node itself) should be updated or deleted.

An example import of content can be:

{
  "id3973361": {
    "comments": {
      "message": {
        "sling:resourceType": "sakai/messagestore",
        "_created": 1334613157826,
        "_id": "e1pKIIgOEeG4WfQRwKgBAw+",
        "_lastModifiedBy": "admin",
        "_lastModified": 1334613157859,
        "_createdBy": "admin",
        "_path": "l6yabccQGq/id9487514/id3973361/comments/message",
        "inbox@Ignore": true
      }
    }
  }
}

Where everything below id3973361/comments/message/inbox, including 'inbox' itself will not be modified or deleted. Note that the following are considered equivalent:

"inbox@Ignore": true
"inbox@Ignore": {}
"inbox@Ignore": []

As the value of the content is not important below this point.