KERN-3045 Added a new FileUploadFilter hook that gets run before storing an uploaded file.
Added a new FileUploadFilter interface that is invoked at the moment just
before an uploaded file is permanently stored in the repository.  Based on the
mime type, request parameters and the byte stream itself, implementers of this
class can return a modified InputStream to be stored in place of the originally
uploaded content.